Improved Object parser
